chore(changesets): :package: version packages (#1)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.1

### Patch Changes

-   [`936f6a5`](https://github.com/kinguru-io/kinguru-monorepo/commit/936f6a5355305a62622e806718eb6a70093164ef) Thanks [@edelwud](https://github.com/edelwud)! - add initial project setup