chore(changesets): :package: version packages (#33)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.3

### Patch Changes

-   Updated dependencies \[[`9d43948`](https://github.com/kinguru-io/kinguru-monorepo/commit/9d439487b03dc43351ff4c9c5ab22836587e005b)]:
    -   database@0.0.1

## database@0.0.1

### Patch Changes

-   [`9d43948`](https://github.com/kinguru-io/kinguru-monorepo/commit/9d439487b03dc43351ff4c9c5ab22836587e005b) Thanks [@edelwud](https://github.com/edelwud)! - add initial database schema