chore(changesets): :package: version packages (#76)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.8

### Patch Changes

-   [#75](https://github.com/kinguru-io/kinguru-monorepo/pull/75) [`38507aa`](https://github.com/kinguru-io/kinguru-monorepo/commit/38507aab27ba4456c14b61a0afc07cb655fa027c) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.3.3`.