chore(changesets): :package: version packages (#78)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.9

### Patch Changes

-   [#77](https://github.com/kinguru-io/kinguru-monorepo/pull/77) [`1d72d48`](https://github.com/kinguru-io/kinguru-monorepo/commit/1d72d48a7b3e8eaa9577500ee84acd7c5e8a544d) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.4.0`.