chore(changesets): :package: version packages (#82)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.11

### Patch Changes

-   [#81](https://github.com/kinguru-io/kinguru-monorepo/pull/81) [`25c6356`](https://github.com/kinguru-io/kinguru-monorepo/commit/25c635621e90d46384417b96c1b26c25b2de1b31) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `stripe` to `14.10.0`.