chore(changesets): :package: version packages (#92)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.15

### Patch Changes

-   [#91](https://github.com/kinguru-io/kinguru-monorepo/pull/91) [`0ed0d05`](https://github.com/kinguru-io/kinguru-monorepo/commit/0ed0d05c3c669857f2bd6d70b1ccf09e4fedc9e7) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `nodemailer` to `6.9.8`.