chore(changesets): :package: version packages (#98)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.17

### Patch Changes

-   [#97](https://github.com/kinguru-io/kinguru-monorepo/pull/97) [`aaf41ae`](https://github.com/kinguru-io/kinguru-monorepo/commit/aaf41ae41dc4d35dd7abd56da5ab85b95a2547be) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.4.1`.