chore(changesets): :package: version packages (#128)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.25

### Patch Changes

-   Updated dependencies \[[`38ca807`](https://github.com/kinguru-io/kinguru-monorepo/commit/38ca80741df6c856190721e214371ba0666d2cad)]:
    -   database@0.0.9

## database@0.0.9

### Patch Changes

-   [#127](https://github.com/kinguru-io/kinguru-monorepo/pull/127) [`38ca807`](https://github.com/kinguru-io/kinguru-monorepo/commit/38ca80741df6c856190721e214371ba0666d2cad) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `prisma` to `5.8.1`.
    Updated dependency `@prisma/client` to `5.8.1`.
    Updated dependency `@prisma/instrumentation` to `5.8.1`.