chore(changesets): :package: version packages (#132)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.26

### Patch Changes

-   [#131](https://github.com/kinguru-io/kinguru-monorepo/pull/131) [`58a4df5`](https://github.com/kinguru-io/kinguru-monorepo/commit/58a4df5907a866288311036a0f356881fd1f1466) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.4.3`.