chore(changesets): :package: version packages (#139)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.29

### Patch Changes

-   [#138](https://github.com/kinguru-io/kinguru-monorepo/pull/138) [`baaded1`](https://github.com/kinguru-io/kinguru-monorepo/commit/baaded19b875086d3033847f596aaae7458b3c4a) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.4.4`.