chore(changesets): :package: version packages (#149)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.33

### Patch Changes

-   [#148](https://github.com/kinguru-io/kinguru-monorepo/pull/148) [`18b28a5`](https://github.com/kinguru-io/kinguru-monorepo/commit/18b28a56464755829135abbb8d84ff082878f476) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `eslint-config-next` to `14.1.0`.