chore(changesets): :package: version packages (#165)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.36

### Patch Changes

-   [#164](https://github.com/kinguru-io/kinguru-monorepo/pull/164) [`20eac9e`](https://github.com/kinguru-io/kinguru-monorepo/commit/20eac9e2d35b60f5a33967f2b377ba62182e3a89) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.5.2`.