chore(changesets): :package: version packages (#167)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.37

### Patch Changes

-   Updated dependencies \[[`7e467ff`](https://github.com/kinguru-io/kinguru-monorepo/commit/7e467ff996b4b67d6280393ba2a5973873d60215)]:
    -   database@0.0.13

## database@0.0.13

### Patch Changes

-   [#166](https://github.com/kinguru-io/kinguru-monorepo/pull/166) [`7e467ff`](https://github.com/kinguru-io/kinguru-monorepo/commit/7e467ff996b4b67d6280393ba2a5973873d60215) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@faker-js/faker` to `8.4.0`.