chore(changesets): :package: version packages (#246)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.41

### Patch Changes

-   [#245](https://github.com/kinguru-io/kinguru-monorepo/pull/245) [`a350ada`](https://github.com/kinguru-io/kinguru-monorepo/commit/a350ada9bc1b3a64b7f5af4a8ffc343940467c41) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `stripe` to `14.21.0`.