chore(changesets): :package: version packages (#248)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.42

### Patch Changes

-   [#247](https://github.com/kinguru-io/kinguru-monorepo/pull/247) [`4c99bdf`](https://github.com/kinguru-io/kinguru-monorepo/commit/4c99bdf0e1bbae8bb9d828079e14c33a9e1c8e3d) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `next-intl` to `3.9.5`.