chore(changesets): :package: version packages (#252)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @kinguru/uikit@0.0.27

### Patch Changes

-   [#251](https://github.com/kinguru-io/kinguru-monorepo/pull/251) [`83f8c1d`](https://github.com/kinguru-io/kinguru-monorepo/commit/83f8c1df7a999c7a41a5acac87bc381a6ee440e8) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@babel/core` to `7.24.1`.
    Updated dependency `@babel/preset-env` to `7.24.1`.
    Updated dependency `@babel/preset-react` to `7.24.1`.
    Updated dependency `@babel/preset-typescript` to `7.24.1`.

## web@0.0.44

### Patch Changes

-   Updated dependencies \[[`83f8c1d`](https://github.com/kinguru-io/kinguru-monorepo/commit/83f8c1df7a999c7a41a5acac87bc381a6ee440e8)]:
    -   @kinguru/uikit@0.0.27