chore(changesets): :package: version packages (#254)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.45

### Patch Changes

-   [#253](https://github.com/kinguru-io/kinguru-monorepo/pull/253) [`02a7d6e`](https://github.com/kinguru-io/kinguru-monorepo/commit/02a7d6e06092b3694541623434215ccb47f31e9b) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `eslint-config-next` to `14.1.4`.
    Updated dependency `next` to `14.1.4`.