chore(changesets): :package: version packages (#256)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## web@0.0.46

### Patch Changes

-   [#255](https://github.com/kinguru-io/kinguru-monorepo/pull/255) [`6618704`](https://github.com/kinguru-io/kinguru-monorepo/commit/6618704d52f4589ecc1afde78a27244bd52bb289) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `nodemailer` to `6.9.13`.