chore(changesets): :package: version packages (#258)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @kinguru/uikit@0.0.28

### Patch Changes

-   [#257](https://github.com/kinguru-io/kinguru-monorepo/pull/257) [`79ba7c2`](https://github.com/kinguru-io/kinguru-monorepo/commit/79ba7c2e730569469396917e346d96d17d48d002) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@babel/core` to `7.24.3`.
    Updated dependency `@babel/preset-env` to `7.24.3`.

## web@0.0.47

### Patch Changes

-   Updated dependencies \[[`79ba7c2`](https://github.com/kinguru-io/kinguru-monorepo/commit/79ba7c2e730569469396917e346d96d17d48d002)]:
    -   @kinguru/uikit@0.0.28